REFACTOR: Moved button into button directory
