+++ components/Filters/FilterPanel/CloseFilterPanelButton.tsx
@@ -1,4 +1,4 @@
-import Button from "@/components/UI/Button";
+import Button from "@/components/Button/Button";
 import React from "react";
 import { MdClear } from "react-icons/md";
